chapter_8-1/requirements.txt
openai
pymongo
langchain-text-splitters
python-dotenv
